Getting locations clusters queries by geohashes within given bounds
A set of geohashes that cover the given bounds is calculated at the zoom
level given. For lower zoom  levels The db is then queries for all locations that fall in these
geohashses, and no MongoDB geospatial query is made.

This will allow for cacheing of the geohashes cluster at the given zoom
level, speeding up the expensive queries at small zoom levels.